tooltip added gallery need some more fix
